+++ packages/headless/src/features/query-suggest/query-suggest-slice.ts
@@ -1,4 +1,6 @@
+import {AnyAction, type Draft as WritableDraft} from '@reduxjs/toolkit';
 import {createReducer} from '@reduxjs/toolkit';
+import {fetchQuerySuggestions as fetchCommerceQuerySuggestions} from '../commerce/query-suggest/query-suggest-actions';
 import {
   clearQuerySuggest,
   fetchQuerySuggestions,
@@ -7,6 +9,7 @@ import {
 } from './query-suggest-actions';
 import {
   getQuerySuggestSetInitialState,
+  QuerySuggestSet,
   QuerySuggestState,
 } from './query-suggest-state';
 
@@ -26,16 +29,7 @@ export const querySuggestReducer = createReducer(
       .addCase(unregisterQuerySuggest, (state, action) => {
         delete state[action.payload.id];
       })
-      .addCase(fetchQuerySuggestions.pending, (state, action) => {
-        const querySuggest = state[action.meta.arg.id];
-
-        if (!querySuggest) {
-          return;
-        }
-
-        querySuggest.currentRequestId = action.meta.requestId;
-        querySuggest.isLoading = true;
-      })
+      .addCase(fetchQuerySuggestions.pending, handleFetchPending)
       .addCase(fetchQuerySuggestions.fulfilled, (state, action) => {
         const querySuggest = state[action.meta.arg.id];
 
@@ -57,16 +51,37 @@ export const querySuggestReducer = createReducer(
         querySuggest.isLoading = false;
         querySuggest.error = null;
       })
-      .addCase(fetchQuerySuggestions.rejected, (state, action) => {
+      .addCase(fetchQuerySuggestions.rejected, handleFetchRejected)
+      .addCase(fetchCommerceQuerySuggestions.pending, handleFetchPending)
+      .addCase(fetchCommerceQuerySuggestions.fulfilled, (state, action) => {
         const querySuggest = state[action.meta.arg.id];
 
-        if (!querySuggest) {
+        if (
+          !querySuggest ||
+          action.meta.requestId !== querySuggest.currentRequestId
+        ) {
           return;
         }
 
-        querySuggest.error = action.payload || null;
+        const {query} = action.payload;
+        if (query) {
+          querySuggest.partialQueries.push(
+            query.replace(/;/, encodeURIComponent(';'))
+          );
+        }
+        querySuggest.responseId = action.payload.responseId;
+        querySuggest.completions = action.payload.completions.map(
+          (completion) => ({
+            expression: completion.expression,
+            highlighted: completion.highlighted,
+            score: 0,
+            executableConfidence: 0,
+          })
+        );
         querySuggest.isLoading = false;
+        querySuggest.error = null;
       })
+      .addCase(fetchCommerceQuerySuggestions.rejected, handleFetchRejected)
       .addCase(clearQuerySuggest, (state, action) => {
         const querySuggest = state[action.payload.id];
 
@@ -94,4 +109,32 @@ function buildQuerySuggest(
     isLoading: false,
     ...config,
   };
+}
+
+function handleFetchPending(
+  state: WritableDraft<QuerySuggestSet>,
+  action: AnyAction
+) {
+  const querySuggest = state[action.meta.arg.id];
+
+  if (!querySuggest) {
+    return;
+  }
+
+  querySuggest.currentRequestId = action.meta.requestId;
+  querySuggest.isLoading = true;
+}
+
+function handleFetchRejected(
+  state: WritableDraft<QuerySuggestSet>,
+  action: AnyAction
+) {
+  const querySuggest = state[action.meta.arg.id];
+
+  if (!querySuggest) {
+    return;
+  }
+
+  querySuggest.error = action.payload || null;
+  querySuggest.isLoading = false;
 }