+++ packages/headless/src/features/search-parameters/search-parameter-analytics-actions.ts
@@ -8,6 +8,7 @@ import {
   facetDeselect,
   facetExclude,
   facetSelect,
+  logFacetUnexclude,
 } from '../../features/facets/facet-set/facet-set-analytics-actions';
 import {
   logSearchboxSubmit,
@@ -110,9 +111,20 @@ function legacyLogFacetAnalyticsAction(
   const removedIds = previousIds.filter((id) => !newIds.includes(id));
   if (removedIds.length) {
     const facetId = removedIds[0];
-    return previousFacets[facetId].length > 1
-      ? logFacetClearAll(facetId)
-      : logFacetDeselect({facetId, facetValue: previousFacets[facetId][0]});
+    switch (true) {
+      case previousFacets[facetId].length > 1:
+        return logFacetClearAll(facetId);
+      case excluded:
+        return logFacetUnexclude({
+          facetId,
+          facetValue: previousFacets[facetId][0],
+        });
+      default:
+        return logFacetDeselect({
+          facetId,
+          facetValue: previousFacets[facetId][0],
+        });
+    }
   }
 
   const addedIds = newIds.filter((id) => !previousIds.includes(id));
@@ -162,10 +174,15 @@ function legacyLogFacetAnalyticsAction(
   );
 
   if (removedValues.length) {
-    return logFacetDeselect({
-      facetId: facetIdWithDifferentValues,
-      facetValue: removedValues[0],
-    });
+    return excluded
+      ? logFacetUnexclude({
+          facetId: facetIdWithDifferentValues,
+          facetValue: removedValues[0],
+        })
+      : logFacetDeselect({
+          facetId: facetIdWithDifferentValues,
+          facetValue: removedValues[0],
+        });
   }
 
   return logInterfaceChange();